==> 01-fl_example_using_mnist_sklearn.ipynb <==
# Federated Learning Case Study

## 01. Example using MNIST

In this notebook we extend the original example of using Flower Federated Learning with scikit-learn (https://flower.dev/docs/quickstart-scikitlearn.html)

* We adapt the online example to make it compatible with a notebook format, using the FL Simulation, over separate client-server Python files.
* We output the individual client accuracy scores to assess their individual and collective performance.
* Minor amendments to enable the example to work within a single Jupyter notebook file.

%%capture
!pip install flwr[simulation] torch torchvision matplotlib

from collections import OrderedDict
from typing import List, Tuple

import flwr as fl
import numpy as np
import matplotlib.pyplot as plt
import torch
import torch.nn as nn
import torchvision
import torch.nn.functional as F
import torchvision.transforms as transforms
from flwr.common import Metrics
from torch.utils.data import DataLoader, random_split
from torchvision.datasets import CIFAR10

import warnings


from typing import Dict

from sklearn.metrics import log_loss

from sklearn.linear_model import LogisticRegression
from sklearn.neural_network import MLPClassifier


print("flwr", fl.__version__)
print("numpy", np.__version__)
print("torch", torch.__version__)
print("torchvision", torchvision.__version__)

DEVICE = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
print(f"Training on {DEVICE}")

from typing import Tuple, Union, List
import numpy as np
from sklearn.linear_model import LogisticRegression
import openml

XY = Tuple[np.ndarray, np.ndarray]
Dataset = Tuple[XY, XY]
LogRegParams = Union[XY, Tuple[np.ndarray]]
XYList = List[XY]

def get_model_parameters(model: LogisticRegression) -> LogRegParams:
    """Returns the paramters of a sklearn LogisticRegression model."""
    if model.fit_intercept:
        params = [
            model.coef_,
            model.intercept_,
        ]
    else:
        params = [
            model.coef_,
        ]
    return params


def set_model_params(
    model: LogisticRegression, params: LogRegParams
) -> LogisticRegression:
    """Sets the parameters of a sklean LogisticRegression model."""
    model.coef_ = params[0]
    if model.fit_intercept:
        model.intercept_ = params[1]
    return model


def set_initial_params(model: LogisticRegression):
    """Sets initial parameters as zeros Required since model params are
    uninitialized until model.fit is called.
    But server asks for initial parameters from clients at launch. Refer
    to sklearn.linear_model.LogisticRegression documentation for more
    information.
    """
    n_classes = 10 # CICIDS has 15 possible classes #10  # MNIST has 10 classes
    n_features = 784 # CICIDS has 78 posible features #784  # Number of features in dataset
    model.classes_ = np.array([i for i in range(10)])

    model.coef_ = np.zeros((n_classes, n_features))
    if model.fit_intercept:
        model.intercept_ = np.zeros((n_classes,))


def load_mnist() -> Dataset:
    """Loads the MNIST dataset using OpenML.
    OpenML dataset link: https://www.openml.org/d/554
    """
    mnist_openml = openml.datasets.get_dataset(554)
    Xy, _, _, _ = mnist_openml.get_data(dataset_format="array")
    X = Xy[:, :-1]  # the last column contains labels
    y = Xy[:, -1]
    # First 60000 samples consist of the train set
    x_train, y_train = X[:60000], y[:60000]
    x_test, y_test = X[60000:], y[60000:]
    return (x_train, y_train), (x_test, y_test)


def shuffle(X: np.ndarray, y: np.ndarray) -> XY:
    """Shuffle X and y."""
    rng = np.random.default_rng()
    idx = rng.permutation(len(X))
    return X[idx], y[idx]


def partition(X: np.ndarray, y: np.ndarray, num_partitions: int) -> XYList:
    """Split X and y into a number of partitions."""
    return list(
        zip(np.array_split(X, num_partitions), np.array_split(y, num_partitions))
    )


NUM_OF_CLIENTS = 10
NUM_OF_ROUNDS = 20

class FlowerClient(fl.client.NumPyClient):
    def __init__(self, cid, c_model, c_X_train, c_y_train, c_X_test, c_y_test):
        self.cid = cid
        self.c_model = c_model
        self.c_X_train = c_X_train
        self.c_y_train = c_y_train
        self.c_X_test = c_X_test
        self.c_y_test = c_y_test

    def get_parameters(self, config):  # type: ignore
        #print (f"Get Parameters for Client {self.cid}")
        return get_model_parameters(self.c_model)

    def fit(self, parameters, config):  # type: ignore
        #print (f"Fit for Client {self.cid}")
        set_model_params(self.c_model, parameters)
        # Ignore convergence failure due to low local epochs
        with warnings.catch_warnings():
            warnings.simplefilter("ignore")
            self.c_model.fit(self.c_X_train, self.c_y_train)
        print(f"Training finished for round {config['server_round']} for Client {self.cid}")
        return get_model_parameters(self.c_model), len(self.c_X_train), {}

    def evaluate(self, parameters, config):  # type: ignore
        #print (f"Evaluate for Client {self.cid}")
        set_model_params(self.c_model, parameters)
        #c_loss = log_loss(self.c_y_test, self.c_model.predict_proba(self.c_X_test))
        #c_accuracy = self.c_model.score(self.c_X_test, self.c_y_test)
        c_loss = log_loss(self.c_y_train, self.c_model.predict_proba(self.c_X_train))
        c_accuracy = self.c_model.score(self.c_X_train, self.c_y_train)
        print (f"Evaluate for Client {self.cid} - Accuracy: {c_accuracy}, Loss: {c_loss}")
        return c_loss, len(self.c_X_train), {"accuracy": c_accuracy}





def client_fn(cid: str) -> FlowerClient:
    """Create a Flower client representing a single organization."""

    # Split train set into 10 partitions and randomly use one for training.
    # Load MNIST dataset from https://www.openml.org/d/554
    (X_train, y_train), (X_test, y_test) = load_mnist()
    partition_id = np.random.choice(10)
    (X_train, y_train) = partition(X_train, y_train, 10)[partition_id]

    # Create LogisticRegression Model
    client_model = LogisticRegression(
        penalty="l2",
        max_iter=1,  # local epoch
        warm_start=True,  # prevent refreshing weights when fitting
    )

    # Setting initial parameters, akin to model.compile for keras models
    set_initial_params(client_model)

    # Create a single Flower client representing a single organization
    return FlowerClient(cid, client_model, X_train, y_train, X_test, y_test)

def fit_round(server_round: int) -> Dict:
    """Send round number to client."""
    return {"server_round": server_round}


def get_evaluate_fn(model: LogisticRegression):
    """Return an evaluation function for server-side evaluation."""

    # Load test data here to avoid the overhead of doing it in `evaluate` itself
    _, (s_X_test, s_y_test) = load_mnist()
    
    # The `evaluate` function will be called after every round
    def evaluate(server_round, parameters: fl.common.NDArrays, config):
        # Update model with the latest parameters
        set_model_params(model, parameters)
        loss = log_loss(s_y_test, model.predict_proba(s_X_test))
        accuracy = model.score(s_X_test, s_y_test)
        return loss, {"accuracy": accuracy}

    return evaluate

server_model = LogisticRegression()
set_initial_params(server_model)

strategy = fl.server.strategy.FedAvg(
    min_available_clients=2,
    evaluate_fn=get_evaluate_fn(server_model),
    on_fit_config_fn=fit_round,
)

fl.simulation.start_simulation(
    client_fn=client_fn,
    num_clients=NUM_OF_CLIENTS,
    strategy=strategy,
    config=fl.server.ServerConfig(num_rounds=NUM_OF_ROUNDS),
    
)
